+++ idea/testData/refactoring/introduceProperty/primaryConstructorParameterReference.kt
@@ -1,0 +1,4 @@
+// EXTRACTION_TARGET: property with initializer
+class Foo(s: String) {
+    val l = <selection>(s + "1")</selection>.length
+}